Refactor solution to smaller modules
